Refactor folder structure
